+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## 0.1.0
+
+Cross-platform support
+* Audio data can be sent from Android to iOS and vice versa.
+* Player that receive stream of PCM 16bit sound data in a (Uint8List).
+* Added new example: Send data via **WebSocket**
+* Fixed some bugs.
+
 ## 0.0.1
 
 Initial release
-* Provides stream of data from mic (Uint8List)
-* Player that receive stream of raw sound data (Uint8List)
-* Support both Android and iOS
+* Provides stream of data from mic (Uint8List).
+* Player that receive stream of raw sound data (Uint8List).
+* Support both Android and iOS.